+++ src/common/tags/tags.cpp
@@ -28,28 +28,6 @@ using namespace libmatroska;
 
 namespace mtx { namespace tags {
 
-void
-fix_mandatory_elements(EbmlElement *e) {
-  if (dynamic_cast<KaxTag *>(e)) {
-    KaxTag &t = *static_cast<KaxTag *>(e);
-    GetChild<KaxTagTargets>(t);
-    GetChild<KaxTagSimple>(t);
-
-  } else if (dynamic_cast<KaxTagSimple *>(e))
-    FixMandatoryElement<KaxTagName, KaxTagLangue, KaxTagDefault>(static_cast<KaxTagSimple *>(e));
-
-  else if (dynamic_cast<KaxTagTargets *>(e)) {
-    KaxTagTargets &t = *static_cast<KaxTagTargets *>(e);
-    GetChild<KaxTagTargetTypeValue>(t);
-    FixMandatoryElement<KaxTagTargetTypeValue>(t);
-
-  }
-
-  if (dynamic_cast<EbmlMaster *>(e))
-    for (auto child : *static_cast<EbmlMaster *>(e))
-      fix_mandatory_elements(child);
-}
-
 KaxTags *
 select_for_chapters(KaxTags &tags,
                     KaxChapters &chapters) {